server/pyproject.toml
[project]
name = "real-state"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "mansuur Abdullahi Abdirahman", email = "mansuurtech101@gmail.com"},
]
dependencies = [
    "django>=5.1.5",
    "djangorestframework>=3.15.2",
    "markdown>=3.7",
    "django-filter>=25.1",
    "pillow>=11.1.0",
    "channels-redis>=4.2.1",
    "uvicorn[standard]>=0.35.0",
    "whitenoise>=6.8.2",
    "python-dotenv>=1.0.1",
    "djangorestframework-simplejwt>=5.4.0",
    "django-ratelimit>=4.1.0",
    "channels>=4.2.0",
    "drf-spectacular[sidecar]>=0.28.0",
    "django-redis>=5.4.0",
    "django-ipware>=7.0.1",
    "django-otp>=1.5.4",
    "django-two-factor-auth>=1.17.0",
    "django-cors-headers>=4.7.0",
    "cryptography>=44.0.0",
    "pendulum>=3.0.0",
    "django-minio-storage>=0.5.7",
    "django-storages[boto3]>=1.14.5",
    "celery[redis]>=5.5.3",
    "pika>=1.3.2",
    "djangorestframework-gis>=1.0",
    "faker>=37.1.0",
    "psycopg2>=2.9.10",
    "django-mptt>=0.17.0",
    "flower>=2.0.1",
    "jinja2>=3.1.6",
    "weasyprint>=65.1",
    "sendgrid>=6.12.4",
    "sentry-sdk[celery,django]>=2.33.2",
    "requests>=2.32.4",
    "djlint>=1.36.4",
    "django-celery-beat>=2.8.1",
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[dependency-groups]
dev = ["ruff>=0.9.9", "black>=25.1.0", "isort>=6.0.1"]


[tool.ruff]
line-length = 88         # Same as Black
target-version = "py311" # Adjust to your Python version

[tool.ruff.lint]
select = ["E", "F", "W", "I"] # Enable common linting rules

[tool.ruff.lint.isort]
known-first-party = ["nhpc-app"]
lines-between-types = 1
combine-as-imports = true

[tool.ruff.lint.pylint]
max-args = 5 # Limit function arguments for cleaner code

[tool.ruff.format]
quote-style = "double" # Enforce double quotes